requirements.txt
uvicorn~=0.22.0
fastapi~=0.100.0
requests~=2.31.0
psycopg2-binary~=2.9.6